% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshared-package.R
\docType{package}
\name{cshared-package}
\alias{cshared}
\alias{cshared-package}
\title{cshared: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
